backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
google-generativeai==0.7.2
groq==0.9.0
tavily-python==0.3.3
sqlmodel==0.0.19
pydantic-settings==2.3.0
python-multipart==0.0.9
reportlab==4.2.2
httpx==0.27.0
pymupdf==1.24.0
python-docx==1.1.0
